+++ src/LivewireModelTable.php
@@ -3,22 +3,18 @@
 namespace timolake\LivewireTables;
 
 use Illuminate\Database\Eloquent\Builder;
-use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\Relations\BelongsToMany;
 use Illuminate\Database\Eloquent\Relations\HasMany;
 use Illuminate\Database\Eloquent\Relations\HasOne;
 use Illuminate\Database\Eloquent\Relations\Relation;
 use Illuminate\Http\Request;
-use Illuminate\Pagination\LengthAwarePaginator;
 use Illuminate\Support\Arr;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Session;
 use Illuminate\Support\Str;
 use Livewire\Component;
 use Livewire\WithPagination;
-use timolake\LivewireTables\Models\Color;
-use timolake\LivewireTables\Models\User;
 
 abstract class LivewireModelTable extends Component
 {
@@ -51,10 +47,6 @@ abstract class LivewireModelTable extends Component
     public array $checkedItems = [];
 
     public string $idField = 'id';
-
-    public array|Collection|LengthAwarePaginator $rowData;
-
-
 
     protected $listeners = ['sortColumn' => 'setSort'];
 
@@ -97,9 +89,9 @@ abstract class LivewireModelTable extends Component
         }
     }
 
-    public function query()
+    protected function query()
     {
-       $this->rowData =  $this->paginate($this->buildQuery());
+        return $this->paginate($this->buildQuery());
     }
 
     protected function buildQuery()