+++ tests/Feature/LivewireTableTest.php
@@ -1,13 +1,11 @@
 <?php
 
 use Illuminate\Foundation\Testing\RefreshDatabase;
-use Illuminate\Support\Facades\View;
 use Livewire\Livewire;
-use timolake\LivewireTables\Models\Color;
-use timolake\LivewireTables\Models\Post;
-use timolake\LivewireTables\Models\Tag;
-use timolake\LivewireTables\Models\User;
-use timolake\LivewireTables\Tests\tables\PostTable;
+use timolake\LivewireTables\Tests\Models\Post;
+use timolake\LivewireTables\Tests\Models\Tag;
+use timolake\LivewireTables\Tests\Models\User;
+use timolake\LivewireTables\Tests\Tables\PostTable;
 
 uses(RefreshDatabase::class);
 
@@ -26,18 +24,27 @@ test("that a table can search in a single table", function () {
 
     Livewire::test(PostTable::class)
         ->set('search', 'foobar')
-        ->call("query")
-        ->assertCount('rowData',5);
+        ->assertSeeText('foobar')
+        ->assertDontSeeText('baz')
+        ->assertDontSeeText('azerty')
+        ->assertViewHas('rowData', function ($rowData) {
+            return count($rowData) == 5;
+        });
 
     Livewire::test(PostTable::class)
         ->set('search', 'baz')
-        ->call("query")
-        ->assertCount('rowData',4);
+        ->assertSeeText('baz')
+        ->assertDontSeeText('foobar')
+        ->assertDontSeeText('azerty')
+        ->assertViewHas('rowData', function ($rowData) {
+            return count($rowData) == 4;
+        });
 });
 
 test("that a table can search in a BelongsTo relation", function () {
 
     $user = User::factory()->create(["name" => "Jefke"]);
+
     Post::factory()->count(5)->create([
         "title" => "foobar",
     ]);
@@ -52,9 +59,14 @@ test("that a table can search in a BelongsTo relation", function () {
     $this->assertDatabaseCount("posts",20);
 
     Livewire::test(PostTable::class)
-        ->set('search', "Jefke")
-        ->call("query")
-        ->assertCount('rowData',11);
+        ->set('search', 'Jefke')
+        ->assertSeeText('Jefke')
+        ->assertSeeText('azerty')
+        ->assertDontSeeText('foobar')
+        ->assertDontSeeText('baz')
+        ->assertViewHas('rowData', function ($rowData) {
+            return count($rowData) == 11;
+        });
 
 });
 
@@ -78,14 +90,25 @@ test("that a table can search in a HasMany relation", function () {
     $this->assertDatabaseCount("posts",16);
 
     Livewire::test(PostTable::class)
-        ->set('search', "foobaz")
-        ->call("query")
-        ->assertCount('rowData',5);
+        ->set('search', 'foobaz')
+        ->assertSeeText('foobaz')
+        ->assertSeeText('qwerty')
+        ->assertDontSeeText('azerty')
+        ->assertDontSeeText('foobar')
+        ->assertViewHas('rowData', function ($rowData) {
+            return count($rowData) == 5;
+        });
 
     Livewire::test(PostTable::class)
-        ->set('search', "baz")
-        ->call("query")
-        ->assertCount('rowData',8);
+        ->set('search', 'baz')
+        ->assertSeeText('baz')
+        ->assertSeeText('foobaz')
+        ->assertSeeText('qwerty')
+        ->assertDontSeeText('azerty')
+        ->assertDontSeeText('foobar')
+        ->assertViewHas('rowData', function ($rowData) {
+            return count($rowData) == 8;
+        });
 
 });
 
@@ -120,20 +143,31 @@ test("that a table can search in a BelongsToMany relation", function () {
     $this->assertDatabaseCount("posts",8);
 
     Livewire::test(PostTable::class)
-        ->set('search', "foo")
-        ->call("query")
-        ->assertCount('rowData',2);
+        ->set('search', 'foo')
+        ->assertSeeText('foo')
+        ->assertSeeText('bar')
+        ->assertDontSeeText('baz')
+        ->assertViewHas('rowData', function ($rowData) {
+            return count($rowData) == 2;
+        });
 
     Livewire::test(PostTable::class)
-        ->set('search', "bar")
-        ->call("query")
-        ->assertCount('rowData',2);
+        ->set('search', 'bar')
+        ->assertSeeText('bar')
+        ->assertSeeText('foo')
+        ->assertDontSeeText('baz')
+        ->assertViewHas('rowData', function ($rowData) {
+            return count($rowData) == 2;
+        });
 
     Livewire::test(PostTable::class)
-        ->set('search', "baz")
-        ->call("query")
-        ->assertCount('rowData',1);
-
+        ->set('search', 'baz')
+        ->assertSeeText('baz')
+        ->assertDontSeeText('foo')
+        ->assertDontSeeText('bar')
+        ->assertViewHas('rowData', function ($rowData) {
+            return count($rowData) == 1;
+        });
 
 });
 
@@ -146,12 +180,15 @@ test("that a table can search in a HasOne relation 3 levels down ", function () 
 
     Post::factory()
         ->count(2)
+
         ->create([
+            "title" => "foobaz",
             "user_id" => $userWithColorYellow->id
         ]);
 
     Post::factory()
         ->count(3)
+        ->withFoobazComment()
         ->create([
             "user_id" => $userWithColorRed->id
         ]);
@@ -166,24 +203,52 @@ test("that a table can search in a HasOne relation 3 levels down ", function () 
     $this->assertDatabaseCount("posts",10);
 
     Livewire::test(PostTable::class)
-        ->set('search', "Yellow jef")
-        ->call("query")
-        ->assertCount('rowData',2);
+        ->set('search', 'Yellow Jef')
+        ->assertSeeText('Jef')
+        ->assertSeeText('Yellow')
+        ->assertDontSeeText('Piet')
+        ->assertDontSeeText('Jan')
+        ->assertViewHas('rowData', function ($rowData) {
+            return count($rowData) == 2;
+        });
 
     Livewire::test(PostTable::class)
-        ->set('search', "Red")
-        ->call("query")
-        ->assertCount('rowData',3);
+        ->set('search', 'Red')
+        ->assertSeeText('Piet')
+        ->assertDontSeeText('Jef')
+        ->assertDontSeeText('Jan')
+        ->assertViewHas('rowData', function ($rowData) {
+            return count($rowData) == 3;
+        });
 
     Livewire::test(PostTable::class)
-        ->set('search', "Blue")
-        ->call("query")
-        ->assertCount('rowData',5);
+        ->set('search', 'Blue')
+        ->assertSeeText('Blue')
+        ->assertSeeText('Jan')
+        ->assertDontSeeText('Jef')
+        ->assertDontSeeText('Piet')
+        ->assertViewHas('rowData', function ($rowData) {
+            return count($rowData) == 5;
+        });
 
     Livewire::test(PostTable::class)
-        ->set('search', "Blue jef")
-        ->call("query")
-        ->assertCount('rowData',0);
+        ->set('search', 'foobaz')
+        ->assertSeeText('foobaz')
+        ->assertDontSeeText('jan')
+        ->assertViewHas('rowData', function ($rowData) {
+            ray($rowData);
+            return count($rowData) == 5;
+        });
+
+    Livewire::test(PostTable::class)
+        ->set('search', 'green marcel')
+        ->assertDontSeeText('Jef')
+        ->assertDontSeeText('Jan')
+        ->assertDontSeeText('Piet')
+        ->assertViewHas('rowData', function ($rowData) {
+            return count($rowData) == 0;
+        });
+
 
 
 });