+++ tests/Models/Color.php
@@ -1,12 +1,10 @@
 <?php
 
-namespace timolake\LivewireTables\Models;
+namespace timolake\LivewireTables\Tests\Models;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
-use Illuminate\Database\Eloquent\Relations\HasMany;
-use Illuminate\Database\Eloquent\Relations\HasOne;
 
 class Color extends Model
 {