+++ tests/database/factories/ColorFactory.php
@@ -3,15 +3,15 @@
 namespace timolake\LivewireTables\Factories;
 
 use Illuminate\Database\Eloquent\Factories\Factory;
-use timolake\LivewireTables\Models\Post;
-use timolake\LivewireTables\Models\User;
+use timolake\LivewireTables\Tests\Models\User;
 
-class TagFactory extends Factory
+class ColorFactory extends Factory
 {
     public function definition(): array
     {
         return [
-            "name" => $this->faker->word
+            "name" => $this->faker->colorName(),
+            "user_id" => user::factory()->create(),
         ];
     }
 }