+++ tests/database/factories/CommentFactory.php
@@ -3,8 +3,7 @@
 namespace timolake\LivewireTables\Factories;
 
 use Illuminate\Database\Eloquent\Factories\Factory;
-use timolake\LivewireTables\Models\Post;
-use timolake\LivewireTables\Models\User;
+use timolake\LivewireTables\Tests\Models\User;
 
 class CommentFactory extends Factory
 {